solana_vote_interface::state::vote_state_v3

Struct VoteState

pub struct VoteState {
    pub node_pubkey: Pubkey,
    pub authorized_withdrawer: Pubkey,
    pub commission: u8,
    pub votes: VecDeque<LandedVote>,
    pub root_slot: Option<Slot>,
    pub authorized_voters: AuthorizedVoters,
    pub prior_voters: CircBuf<(Pubkey, Epoch, Epoch)>,
    pub epoch_credits: Vec<(Epoch, u64, u64)>,
    pub last_timestamp: BlockTimestamp,
}

Fields
node_pubkey: Pubkey

the node that votes in this account

authorized_withdrawer: Pubkey

the signer for withdrawals

commission: u8

percentage (0-100) that represents what part of a rewards payout should be given to this VoteAccount

votes: VecDeque<LandedVote>root_slot: Option<Slot>authorized_voters: AuthorizedVoters

the signer for vote transactions

prior_voters: CircBuf<(Pubkey, Epoch, Epoch)>

history of prior authorized voters and the epochs for which they were set, the bottom end of the range is inclusive, the top of the range is exclusive

epoch_credits: Vec<(Epoch, u64, u64)>

history of how many credits earned by the end of each epoch each tuple is (Epoch, credits, prev_credits)

last_timestamp: BlockTimestamp

most recent timestamp submitted with a vote

Implementations

impl VoteState

pub fn new(vote_init: &VoteInit, clock: &Clock) -> Self

pub fn new_rand_for_tests(node_pubkey: Pubkey, root_slot: Slot) -> Self

pub fn get_authorized_voter(&self, epoch: Epoch) -> Option<Pubkey>

pub fn authorized_voters(&self) -> &AuthorizedVoters

pub fn prior_voters(&mut self) -> &CircBuf<(Pubkey, Epoch, Epoch)>

pub fn get_rent_exempt_reserve(rent: &Rent) -> u64

pub const fn size_of() -> usize

Upper limit on the size of the Vote State when votes.len() is MAX_LOCKOUT_HISTORY.

pub fn deserialize(input: &[u8]) -> Result<Self, InstructionError>

pub fn deserialize_into( input: &[u8], vote_state: &mut VoteState, ) -> Result<(), InstructionError>

Deserializes the input VoteStateVersions buffer directly into the provided VoteState.

In a SBPF context, V0_23_5 is not supported, but in non-SBPF, all versions are supported for compatibility with bincode::deserialize.

On success, vote_state reflects the state of the input data. On failure, vote_state is reset to VoteState::default().

pub fn deserialize_into_uninit( input: &[u8], vote_state: &mut MaybeUninit<VoteState>, ) -> Result<(), InstructionError>

Deserializes the input VoteStateVersions buffer directly into the provided MaybeUninit<VoteState>.

In a SBPF context, V0_23_5 is not supported, but in non-SBPF, all versions are supported for compatibility with bincode::deserialize.

On success, vote_state is fully initialized and can be converted to VoteState using [MaybeUninit::assume_init]. On failure, vote_state may still be uninitialized and must not be converted to VoteState.

pub fn serialize( versioned: &VoteStateVersions, output: &mut [u8], ) -> Result<(), InstructionError>

pub fn commission_split(&self, on: u64) -> (u64, u64, bool)

๐Deprecated since 2.2.0: logic was moved into the agave runtime crate

returns commission split as (voter_portion, staker_portion, was_split) tuple

if commission calculation is 100% one way or other, indicate with false for was_split

pub fn contains_slot(&self, candidate_slot: Slot) -> bool

Returns if the vote state contains a slot candidate_slot

pub fn process_next_vote_slot( &mut self, next_vote_slot: Slot, epoch: Epoch, current_slot: Slot, )

pub fn increment_credits(&mut self, epoch: Epoch, credits: u64)

increment credits, record credits for last epoch if new epoch

pub fn compute_vote_latency(voted_for_slot: Slot, current_slot: Slot) -> u8

pub fn credits_for_vote_at_index(&self, index: usize) -> u64

Returns the credits to award for a vote at the given lockout slot index

pub fn nth_recent_lockout(&self, position: usize) -> Option<&Lockout>

pub fn last_lockout(&self) -> Option<&Lockout>

pub fn last_voted_slot(&self) -> Option<Slot>

pub fn tower(&self) -> Vec<Slot>

pub fn current_epoch(&self) -> Epoch

pub fn credits(&self) -> u64

Number of โcreditsโ owed to this account from the mining pool. Submit this VoteState to the Rewards program to trade credits for lamports.

pub fn epoch_credits(&self) -> &Vec<(Epoch, u64, u64)>

Number of โcreditsโ owed to this account from the mining pool on a per-epoch basis, starting from credits observed. Each tuple of (Epoch, u64, u64) is read as (epoch, credits, prev_credits), where credits for each epoch is credits - prev_credits; while redundant this makes calculating rewards over partial epochs nice and simple

pub fn set_new_authorized_voter<F>( &mut self, authorized_pubkey: &Pubkey, current_epoch: Epoch, target_epoch: Epoch, verify: F, ) -> Result<(), InstructionError>
where F: Fn(Pubkey) -> Result<(), InstructionError>,

pub fn get_and_update_authorized_voter( &mut self, current_epoch: Epoch, ) -> Result<Pubkey, InstructionError>

pub fn pop_expired_votes(&mut self, next_vote_slot: Slot)

pub fn double_lockouts(&mut self)

pub fn process_timestamp( &mut self, slot: Slot, timestamp: UnixTimestamp, ) -> Result<(), VoteError>

pub fn is_correct_size_and_initialized(data: &[u8]) -> bool